Reorganized the types into constant folder.
