+++ src/backend/KnowledgeNetwork.Core/Models/Constants/EdgeCategories.cs
@@ -1,0 +1,14 @@
+namespace KnowledgeNetwork.Core.Models.Constants;
+
+/// <summary>
+/// Categories for edge types
+/// </summary>
+public static class EdgeCategories
+{
+    public const string Structure = "structure";
+    public const string Execution = "execution";
+    public const string ControlFlow = "control-flow";
+    public const string Dependency = "dependency";
+    public const string Documentation = "documentation";
+    public const string View = "view";
+}
