+++ src/backend/KnowledgeNetwork.Core/Models/Constants/EdgeTypes.cs
@@ -33,16 +33,3 @@ public static class EdgeTypes
     public const string Aggregates = "aggregates";
     public const string FocusesOn = "focuses-on";
 }
-
-/// <summary>
-/// Categories for edge types
-/// </summary>
-public static class EdgeCategories
-{
-    public const string Structure = "structure";
-    public const string Execution = "execution";
-    public const string ControlFlow = "control-flow";
-    public const string Dependency = "dependency";
-    public const string Documentation = "documentation";
-    public const string View = "view";
-}
